Fix merging for real and show multi-pass merging
The sanity checks were failing when another merge was applied. I
realized that the sanity checker should check that for every edge from A
to B there must be an edge from B to A.

Now the sanity checks were failing from the beginning and the issue was
quickly identified.